requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # For PostgreSQL
pytest==7.4.0
pytest-cov==4.1.0